requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
mega.py==1.0.8
pathlib==1.0.1
pycryptodome==3.21.0
pynput==1.7.7
requests==2.32.3
six==1.16.0
tenacity==9.0.0
urllib3==2.2.3
